feat(websocket): add role-based mapping for clients
Add `clientsByRole` property to `Hub` that maps a `user.Role` to the
clients with that role.

This makes it a lot more efficient to send messages to clients with
specific roles with the help of `Hub.sendToRole` method rather than
iterating over every client and checking their roles one by one.