fix(timer): start `handleTimer` go routine before running `startTimer`
`handleTimer` is the method that reads from the `timer.start` and
`timer.done` channels. Running `startTimer` before `handleTimer` causes
the program to be stuck forever since nothing is reading from the
`start` and `end` channels.

This commit fixes that issue, but still has some issues such as a race
condition where `handleTimer` attempts to access a timer that hasn't
been initialized.